Separate IContextMenuUIActionInterface from its implementation.
