+++ Sandra.UI/MovesTextBox.UIActions.cs
@@ -21,7 +21,7 @@
 
 using Eutherion.UIActions;
 using Eutherion.Utils;
-using Eutherion.Win.UIActions;
+using Eutherion.Win.AppTemplate;
 
 namespace Sandra.UI
 {
@@ -33,7 +33,7 @@ namespace Sandra.UI
             new UIAction(MovesTextBoxUIActionPrefix + nameof(UsePGNPieceSymbols)),
             new ImplementationSet<IUIActionInterface>
             {
-                new ContextMenuUIActionInterface
+                new CombinedUIActionInterface
                 {
                     IsFirstInGroup = true,
                     MenuCaptionKey = LocalizedStringKeys.UsePGNPieceSymbols,
@@ -59,7 +59,7 @@ namespace Sandra.UI
             new UIAction(MovesTextBoxUIActionPrefix + nameof(UseLongAlgebraicNotation)),
             new ImplementationSet<IUIActionInterface>
             {
-                new ContextMenuUIActionInterface
+                new CombinedUIActionInterface
                 {
                     MenuCaptionKey = LocalizedStringKeys.UseLongAlgebraicNotation,
                 },
